+++ bot.py
@@ -25,6 +25,7 @@ intents.message_content = True
 bot = commands.Bot(command_prefix='!',intents=intents)
 
 async def load_cogs():
+    
     try:
         for filename in os.listdir('./cogs'):
             if filename.endswith('.py') and filename != '__init__.py':