+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.6.0]
+### Added
+- **Weekly Activity Tracking**
+  - Every Sunday, find out who spent the most time on voice chat!
+### Fixes
+- New alias for K'Sante
 ## [0.5.3] - 21-11-2025
 ### Fixes
 - Multiple new aliases added, focused on making the detection of champions with doubled letters or *"y"*s in their name more indulgent.
@@ -24,4 +30,3 @@
   - `!register` command for individual registration.
 
 - Remember to use  `!help` to see all the commands!
-